All tests pass again.
--HG--
branch : 0.7
