+++ tests/test_ext_jinja2.py
@@ -9,6 +9,7 @@ import unittest
 from jinja2 import FileSystemLoader, Environment
 
 from tipfy import RequestHandler, Request, Response, Tipfy
+from tipfy.app import local
 from tipfyext.jinja2 import Jinja2, Jinja2Mixin
 
 current_dir = os.path.abspath(os.path.dirname(__file__))
@@ -18,15 +19,16 @@ templates_compiled_target = os.path.join(current_dir, 'resources', 'templates_co
 
 class TestJinja2(unittest.TestCase):
     def tearDown(self):
-        Tipfy.app.clear_locals()
+        local.__release_local__()
 
     def test_render_template(self):
         app = Tipfy(config={'tipfyext.jinja2': {'templates_dir': templates_dir}})
-        app.set_locals(Request.from_values())
+        request = Request.from_values()
+        local.current_handler = handler = RequestHandler(app, request)
         jinja2 = Jinja2(app)
 
         message = 'Hello, World!'
-        res = jinja2.render_template('template1.html', message=message)
+        res = jinja2.render_template(local.current_handler, 'template1.html', message=message)
         self.assertEqual(res, message)
 
     def test_render_template_with_i18n(self):
@@ -42,20 +44,22 @@ class TestJinja2(unittest.TestCase):
                 'secret_key': 'secret',
             },
         })
-        app.set_locals(Request.from_values())
+        request = Request.from_values()
+        local.current_handler = handler = RequestHandler(app, request)
         jinja2 = Jinja2(app)
 
         message = 'Hello, i18n World!'
-        res = jinja2.render_template('template2.html', message=message)
+        res = jinja2.render_template(local.current_handler, 'template2.html', message=message)
         self.assertEqual(res, message)
 
     def test_render_response(self):
         app = Tipfy(config={'tipfyext.jinja2': {'templates_dir': templates_dir}})
-        app.set_locals(Request.from_values())
+        request = Request.from_values()
+        local.current_handler = handler = RequestHandler(app, request)
         jinja2 = Jinja2(app)
 
         message = 'Hello, World!'
-        response = jinja2.render_response('template1.html', message=message)
+        response = jinja2.render_response(local.current_handler, 'template1.html', message=message)
         self.assertEqual(isinstance(response, Response), True)
         self.assertEqual(response.mimetype, 'text/html')
         self.assertEqual(response.data, message)
@@ -67,11 +71,12 @@ class TestJinja2(unittest.TestCase):
                 'force_use_compiled': True,
             }
         }, debug=False)
-        app.set_locals(Request.from_values())
+        request = Request.from_values()
+        local.current_handler = handler = RequestHandler(app, request)
         jinja2 = Jinja2(app)
 
         message = 'Hello, World!'
-        response = jinja2.render_response('template1.html', message=message)
+        response = jinja2.render_response(local.current_handler, 'template1.html', message=message)
         self.assertEqual(isinstance(response, Response), True)
         self.assertEqual(response.mimetype, 'text/html')
         self.assertEqual(response.data, message)
@@ -84,11 +89,11 @@ class TestJinja2(unittest.TestCase):
                 self.context = {}
 
         app = Tipfy(config={'tipfyext.jinja2': {'templates_dir': templates_dir}})
-        app.set_locals(Request.from_values())
+        request = Request.from_values()
+        local.current_handler = handler = MyHandler(app, request)
         jinja2 = Jinja2(app)
         message = 'Hello, World!'
 
-        handler = MyHandler(Tipfy.app, Tipfy.request)
         response = handler.render_template('template1.html', message=message)
         self.assertEqual(response, message)
 
@@ -100,11 +105,11 @@ class TestJinja2(unittest.TestCase):
                 self.context = {}
 
         app = Tipfy(config={'tipfyext.jinja2': {'templates_dir': templates_dir}})
-        app.set_locals(Request.from_values())
+        request = Request.from_values()
+        local.current_handler = handler = MyHandler(app, request)
         jinja2 = Jinja2(app)
         message = 'Hello, World!'
 
-        handler = MyHandler(Tipfy.app, Tipfy.request)
         response = handler.render_response('template1.html', message=message)
         self.assertEqual(isinstance(response, Response), True)
         self.assertEqual(response.mimetype, 'text/html')
@@ -112,7 +117,8 @@ class TestJinja2(unittest.TestCase):
 
     def test_get_template_attribute(self):
         app = Tipfy(config={'tipfyext.jinja2': {'templates_dir': templates_dir}})
-        app.set_locals(Request.from_values())
+        request = Request.from_values()
+        local.current_handler = handler = RequestHandler(app, request)
         jinja2 = Jinja2(app)
 
         hello = jinja2.get_template_attribute('hello.html', 'hello')
@@ -120,7 +126,7 @@ class TestJinja2(unittest.TestCase):
 
     def test_engine_factory(self):
         def get_jinja2_env():
-            app = Tipfy.app
+            app = local.current_handler.app
             cfg = app.get_config('tipfyext.jinja2')
 
             loader = FileSystemLoader(cfg.get( 'templates_dir'))
@@ -131,11 +137,12 @@ class TestJinja2(unittest.TestCase):
             'templates_dir': templates_dir,
             'engine_factory': get_jinja2_env,
         }})
-        app.set_locals(Request.from_values())
+        request = Request.from_values()
+        local.current_handler = handler = RequestHandler(app, request)
         jinja2 = Jinja2(app)
 
         message = 'Hello, World!'
-        res = jinja2.render_template('template1.html', message=message)
+        res = jinja2.render_template(local.current_handler, 'template1.html', message=message)
         self.assertEqual(res, message)
 
     def test_engine_factory2(self):
@@ -146,18 +153,20 @@ class TestJinja2(unittest.TestCase):
             'templates_dir': templates_dir,
             'engine_factory': 'resources.get_jinja2_env',
         }})
-        app.set_locals(Request.from_values())
+        request = Request.from_values()
+        local.current_handler = handler = RequestHandler(app, request)
         jinja2 = Jinja2(app)
 
         message = 'Hello, World!'
-        res = jinja2.render_template('template1.html', message=message)
+        res = jinja2.render_template(local.current_handler, 'template1.html', message=message)
         self.assertEqual(res, message)
 
         sys.path = old_sys_path
 
     def test_engine_factory3(self):
         app = Tipfy()
-        app.set_locals(Request.from_values())
+        request = Request.from_values()
+        local.current_handler = handler = RequestHandler(app, request)
         _globals = {'message': 'Hey there!'}
         filters = {'ho': lambda e: e + ' Ho!'}
         jinja2 = Jinja2(app, _globals=_globals, filters=filters)
