+++ tests/test_ext_xmpp.py
@@ -7,6 +7,7 @@ import sys
 import unittest
 
 from tipfy import Rule, Tipfy
+from tipfy.app import local
 
 from google.appengine.api.xmpp import Message as ApiMessage
 
@@ -75,10 +76,7 @@ class TestCommandHandler(unittest.TestCase):
         xmpp.xmpp = sys.modules[__name__]
 
     def tearDown(self):
-        try:
-            Tipfy.app.clear_locals()
-        except:
-            pass
+        local.__release_local__()
         fake_local.clear()
 
         from tipfyext.appengine import xmpp
